updated code to mirro what happened during livecoding  and added presentation schema used during presentation
