+++ components/presentation.jsx
@@ -3,8 +3,8 @@ import _ from 'lodash';
 import Agenda from './agenda';
 import Slide from './slide';
 import Switcher from './switcher';
-import KeyboardSwitcher from './KeyboardSwitcher';
-import * as presentationData from '../theRealDeal';
+import KeyboardSwitcher from './keyboardSwitcher';
+import * as presentationData from '../presentationSchema';
 
 let Presentation = React.createClass({
 
@@ -46,9 +46,9 @@ let Presentation = React.createClass({
                 />
 
                 <Switcher
-                    availableSlidesNumbers={_.range(presentationData.slides.length)}
                     currentSlideNumber={this.state.currentSlideNumber}
                     handleSlideChange={this._handleSlideChange}
+                    slidesCount={presentationData.slides.length}
                     toggleAgenda={this._toggleAgenda}
                 />
