+++ components/slide.css
@@ -5,10 +5,11 @@
     height: 100%;
     text-align: center;
     width: 100%;
+    font-size:2rem;
 }
 
 .Slide h1 {
-    font-size: 4em;
+    font-size: 3em;
     margin: 20px;
 }
 
@@ -17,7 +18,7 @@
     margin: 20px;
 }
 
-.Slide img{
+.Slide img {
     max-height: 500px;
 }
