+++ package.json
@@ -23,7 +23,6 @@
     "classnames": "^2.1.3",
     "css-loader": "^0.18.0",
     "file-loader": "^0.8.4",
-    "lodash": "^3.10.1",
     "react": "^0.13.3",
     "react-hot-loader": "^1.3.0",
     "style-loader": "^0.12.3",
